requirements.txt
fastapi
uvicorn
python-dotenv
cryptography
PyMuPDF
python-docx
pandas
whoosh
openai
faiss-cpu